MSRZ*: remove eth1 and adapt RGMII phy mode and clock skews
Ethernet does work with Gigabit but not with Fast Ethernet.
It works fine in Linux. The primary problem is, that the ravb
network driver can not handle properly two interfaces. There
is only one instance of the mii bitbanging device. As a result,
the mii access to the phy does not work any longer, even for
the first device. For the time being we fix it by removing
(uncommenting) the secoond ethernet interface (eth1) in the DTS
file.

We also overtake the phy settings, rgmii mode "rgmii-id" and
the clock skews from Linux.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>